requirements.txt
mysql-connector-python==8.0.24
contextlib2==0.6.0.post1
fuzzywuzzy==0.18.0